% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wv_and_datasheet.R
\name{plot_gmwm_and_datasheet}
\alias{plot_gmwm_and_datasheet}
\title{Plot GMWM with datasheet characteristics}
\usage{
plot_gmwm_and_datasheet(object, datasheet, prov_tile = NULL)
}
\arguments{
\item{object}{A \code{gmwm} object.}

\item{datasheet}{A \code{vector} with the implied WV of the datasheet.}

\item{prov_tile}{A \code{string} for the title of the graph.}
}
\description{
Plot empirical WV of a sensors together the empirical WV implied by
the datasheet of the manufacture as well as GMWM solution
}
\examples{
Zt = rnorm(1000000)
wv = gmwm(WN(), Zt, freq = 500)
datasheet = wn_to_wv(sigma2 = 1, tau = wv$scales)
plot_gmwm_and_datasheet(wv, datasheet/500)
}
\author{
Stephane Guerrier
}
