% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wv_and_datasheet.R
\name{plot_wv_and_datasheet}
\alias{plot_wv_and_datasheet}
\title{Plot WV with datasheet characteristics}
\usage{
plot_wv_and_datasheet(wv, datasheet, axis.x.label = expression(paste("Scale ",
  tau)), prov_tile = NA)
}
\arguments{
\item{datasheet}{A \code{vector} with the implied WV of the datasheet.}

\item{axis.x.label}{A \code{string} for the label of the "x-axis".}

\item{prov_tile}{A \code{string} for the title of the graph.}

\item{object}{A \code{gmwm} object.}
}
\description{
Plot empirical WV of a sensors together the empirical WV implied by
the datasheet of the manufacture
}
\examples{
Zt = rnorm(1000000)
wv = wvar(Zt, freq = 500)
datasheet = wn_to_wv(sigma2 = 1, tau = wv$scales)
plot_wv_and_datasheet(wv, datasheet/500)
}
\author{
Stephane Guerrier
}
